[NETTOYAGE] extrait l'adaptateur vue/vue.js commun
