+++ src/app/situation_securite.js
@@ -4,9 +4,9 @@ import { afficheSituation } from 'commun/vues/affiche_situation';
 
 import DepotRessourcesSecurite from 'securite/infra/depot_ressources_securite';
 import Situation from 'commun/modeles/situation';
-import VueSituation from 'securite/vues/situation';
+import AdaptateurVueSituation from 'securite/vues/situation';
 
 const situation = new Situation({ modeEntrainement: true, aideDisponible: true });
 
 const depotRessources = new DepotRessourcesSecurite();
-afficheSituation('securite', situation, VueSituation, depotRessources);
+afficheSituation('securite', situation, AdaptateurVueSituation, depotRessources);
